# Speed up testing using parallelization