--- a/build.gradle
+++ b/build.gradle
@@ -66,7 +66,7 @@ dependencies {
 
 tasks.withType(Test).configureEach {
     useJUnitPlatform()
-
+    maxParallelForks = Runtime.runtime.availableProcessors().intdiv(2) ?: 1
     maxHeapSize = '1G'
 }