- [x] Use parallel execution for test
    - Used basic implementation from [Gradle Docs](https://docs.gradle.org/current/userguide/performance.html#execute_tests_in_parallel) as a start
    - Tests seem already properly independent and hence parallelizable

CI effect: 
- before: `+/- 3m 30s` per action (total usage for 5 actions: `15m 31s`)
- after: `2m 58s` per action (total usage for 5 actions: `13m 36s`)

Local effect: On my device (Apple M1 Max, 64GB Memory) running `gradle clean test`:
- before: ` 1m 42s`
- after: `55s`